+++ src/backend/semantic/types.rs
@@ -4,7 +4,7 @@ pub(super) const BUILTIN_EXPRESSION_SIGNATURE: &str = "\\\\expression";
 pub(super) const BUILTIN_STATEMENT_SIGNATURE: &str = "\\\\statement";
 pub(super) const BUILTIN_SPECIFICATION_SIGNATURE: &str = "\\\\specification";
 pub(super) const BUILTIN_TYPE_SIGNATURE: &str = "\\\\type";
-pub(super) const BUILTIN_UNKNOWN_SIGNATURE: &str = "\\\\unknown";
+pub(super) const BUILTIN_OPAQUE_SIGNATURE: &str = "\\\\opaque";
 
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub(super) struct SignatureShape {